requirements.txt
streamlit==1.45.1
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.13.0
requests>=2.28.0
python-dotenv>=0.19.0
openai>=0.27.0
matplotlib==3.7.1
google-cloud-bigquery>=3.3.0
pandas-gbq==0.19.2
google-auth==2.27.0
gspread==6.0.2
google-api-python-client==2.118.0
oauth2client==4.1.3
gspread_dataframe==3.3.1
db-dtypes==1.2.0
networkx==3.2.1
pyvis==0.3.2
cachetools>=5.0.0